% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{app_server}
\alias{app_server}
\title{Server logic for the Shiny application}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object containing input values from the UI.}

\item{output}{Shiny output object used to send output to the UI.}

\item{session}{Shiny session object.}
}
\value{
None. This function is called for its side effects of setting up the server logic of the Shiny app.
}
\description{
This function defines the server-side logic of the Shiny application. It handles user inputs,
processes data, and updates outputs accordingly. The server function sets up reactive expressions,
observers, and render functions to interact with the UI components.
}
\details{
The server function performs several key operations:
\describe{
\item{\strong{Package Validation}}{Validates the selected package name and checks if it is a valid CRAN package.}
\item{\strong{Version Handling}}{Retrieves package versions and updates version selection inputs in the UI.}
\item{\strong{NEWS Diff Tab}}{Fetches the NEWS file for the package and displays the differences between versions.}
\item{\strong{Downloads Tab}}{Plots the download statistics of the package over time.}
\item{\strong{Checkpage Tab}}{Displays CRAN check results and package status.}
\item{\strong{Release List Tab}}{Shows the list of package releases with details.}
\item{\strong{Dependencies Tab}}{Displays user and developer dependencies of the package.}
\item{\strong{Info Tab}}{Provides package information such as title, description, version, license, etc.}
}
}
