% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{app_ui}
\alias{app_ui}
\title{User Interface for the Shiny Application}
\usage{
app_ui()
}
\value{
A Shiny UI definition created using \code{shinydashboard::dashboardPage}.
}
\description{
This function defines the user interface of the Shiny dashboard application. It sets up the dashboard layout,
including the header, sidebar, and body with multiple tabs for different functionalities such as package information,
downloads, version differences, CRAN check results, dependencies, releases, and about.
}
