% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rdevdash_app.R
\name{run_rdevdash_app}
\alias{run_rdevdash_app}
\title{Run the R Dev Dashboard Shiny Application}
\usage{
run_rdevdash_app()
}
\value{
A Shiny app object that represents the running application.
}
\description{
This function launches the R Dev Dashboard Shiny application by initializing the user interface and server logic
defined in \code{app_ui} and \code{app_server}. It starts the application in your default web browser.
}
\examples{
\dontrun{
if (interactive()) {
  run_rdevdash_app()
}
}
}
